ble/requirements.txt
dbus-python
omegaconf
pycairo
PyGObject